make MAX_NOTES int

+++ src/Constants.h
@@ -20,7 +20,7 @@ namespace constants {
         StretchVoiceStepsPattern
     };
 
-    constexpr size_t MAX_NOTES = 24;
+    constexpr int MAX_NOTES = 24;
     constexpr stepnum_t MAX_STEPS = 16;
     constexpr voicenum_t MAX_VOICES = 8;
     constexpr int MAX_OCTAVES = 4;
